separated VC tests
